Update 1 on experimental
This is a HUGE update on Pagination (lazy loading). Now when you open a chat you are going to load certain amount of messages (for now 20), and only if you will scroll up it will load more, and then more, and more... This is a great optimization considering checking for new messages is also optimized.