Update 2 on experimental
To make messaging more safe, I added a cooldown on sending messages. Now you will need to wait at least 1 second before sending a new message. It's both client-side and server-side (in case of abuse). Also now ONLY if you are a sender the messages will automatically be scrolled to the sended one. (It will happen to a reciever too if you are close to that new message).